remove deprecated code and unused files
